pyproject.toml
[tool.poetry]
name = "tevildo"
version = "0.1.0"
description = "A python based OpenAI bot for shell environments"
authors = ["Amarpreet Singh <solamarpreet@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.3.7"
prompt-toolkit = "^3.0.41"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
tevildo = "tevildo.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
